Added working out of exercise 1.19
The solution for this exercise currently only shows the resulting function,
without working out the souble application of the transformation. It's important
for the student to understand not only how to apply the transformation twice,
but to see how to rearrange the terms so the values of p' and q' emerges.